# Vectorizing vector_concat for improved performance