--- a/src/vector.c
+++ b/src/vector.c
@@ -916,12 +916,11 @@ vector_concat(PG_FUNCTION_ARGS)
 	Vector	   *b = PG_GETARG_VECTOR_P(1);
 	Vector	   *result;
 	int			dim = a->dim + b->dim;
+	const int dim_a = a->dim;
+	const int dim_b = b->dim;
 
 	CheckDim(dim);
 	result = InitVector(dim);
-
-	const int dim_a = a->dim;
-	const int dim_b = b->dim;
 
 	/* Auto-vectorized */
 	for (int i = 0; i < dim_a; i++)