@jkatz 2025-06-17
@binarycleric Do you have any before/after benchmarks on this? There's some [examples of benchmarks to run and what to test here](https://jkatz05.com/post/postgres/pgvector-performance-150x-speedup/).