@binarycleric 2025-06-17
The following was generated using https://github.com/binarycleric/pgvectorbench/blob/main/benchmarks/checks/vector_concat.sql and tested against Postgres 16 on macOS with M4 Pro chip. I'm hitting EoD right now but tomorrow I'm going to run these tests on r7i/r7g instances and post results.

## With Change

Source from the `vectorize-vector-concat` branch.

```
 function_name |     test_name      | Minimum Time (us) | Median Time (us) | 95th Percentile (us) | 99th Percentile (us) | Maximum Time (us) | Standard Deviation
---------------+--------------------+-------------------+------------------+----------------------+----------------------+-------------------+--------------------
 vector_concat | small_vectors      |             9.775 |           10.967 |               13.113 |               14.067 |           663.996 |              6.610
 vector_concat | medium_vectors     |            12.875 |           14.067 |               16.928 |               17.166 |           116.110 |              1.647
 vector_concat | large_vectors      |           191.927 |          204.086 |              224.829 |              254.154 |           506.878 |             12.241
 vector_concat | very_large_vectors |           623.941 |          651.121 |              739.813 |              777.006 |          1362.085 |             38.261
 ```

 ## Without Change

Latest source from the `master` branch.

 ```
 function_name |     test_name      | Minimum Time (us) | Median Time (us) | 95th Percentile (us) | 99th Percentile (us) | Maximum Time (us) | Standard Deviation
---------------+--------------------+-------------------+------------------+----------------------+----------------------+-------------------+--------------------
 vector_concat | small_vectors      |            10.967 |           11.921 |               14.067 |               15.974 |           658.989 |              6.564
 vector_concat | medium_vectors     |            18.835 |           20.027 |               22.888 |               25.034 |           118.017 |              1.717
 vector_concat | large_vectors      |           238.895 |          256.777 |              273.943 |              280.857 |           566.006 |             10.689
 vector_concat | very_large_vectors |           768.185 |          813.007 |              890.970 |              950.098 |          1585.007 |             42.797
 ```