@binarycleric 2025-06-18
Here are some more benchmarks, this time on EC2.

## x86_64

Instance Type: r7i.2xlarge
Version: PostgreSQL 16.9

### main branch

```
 function_name |     test_name      | Minimum Time (us) | Median Time (us) | 95th Percentile (us) | 99th Percentile (us) | Maximum Time (us) | Standard Deviation
---------------+--------------------+-------------------+------------------+----------------------+----------------------+-------------------+--------------------
 vector_concat | small_vectors      |            24.796 |           27.180 |               31.948 |               36.955 |           611.067 |              6.256
 vector_concat | medium_vectors     |            31.948 |           35.048 |               41.962 |               46.015 |           216.007 |              3.704
 vector_concat | large_vectors      |           617.981 |          658.989 |              724.077 |              749.111 |          1165.867 |             30.209
 vector_concat | very_large_vectors |          1962.900 |         2108.097 |             2403.021 |             2696.991 |          3768.921 |            154.491
```

### vectorize-vector-concat branch

```
 function_name |     test_name      | Minimum Time (us) | Median Time (us) | 95th Percentile (us) | 99th Percentile (us) | Maximum Time (us) | Standard Deviation
---------------+--------------------+-------------------+------------------+----------------------+----------------------+-------------------+--------------------
 vector_concat | small_vectors      |            23.842 |           26.941 |               30.041 |               37.909 |           607.967 |              6.420
 vector_concat | medium_vectors     |            20.027 |           23.127 |               26.941 |               34.094 |           205.040 |              3.094
 vector_concat | large_vectors      |           521.898 |          581.026 |              638.008 |              656.843 |          1040.936 |             30.592
 vector_concat | very_large_vectors |          1616.001 |         1713.991 |             1950.979 |             2032.995 |          3578.186 |             93.572
```

## ARM

Instance Type: r7g.2xlarge
Version: PostgreSQL 16.9

### main branch

```
 function_name |     test_name      | Minimum Time (us) | Median Time (us) | 95th Percentile (us) | 99th Percentile (us) | Maximum Time (us) | Standard Deviation
---------------+--------------------+-------------------+------------------+----------------------+----------------------+-------------------+--------------------
 vector_concat | small_vectors      |            32.902 |           34.094 |               35.048 |               41.008 |           663.042 |              6.480
 vector_concat | medium_vectors     |            51.975 |           52.929 |               61.035 |               64.135 |           243.902 |              3.406
 vector_concat | large_vectors      |           901.937 |          937.939 |              962.973 |              970.125 |          1309.872 |             14.930
 vector_concat | very_large_vectors |          2517.939 |         2556.086 |             2580.881 |             2666.950 |          3643.990 |             24.754
```

### vectorize-vector-concat branch

```
 function_name |     test_name      | Minimum Time (us) | Median Time (us) | 95th Percentile (us) | 99th Percentile (us) | Maximum Time (us) | Standard Deviation
---------------+--------------------+-------------------+------------------+----------------------+----------------------+-------------------+--------------------
 vector_concat | small_vectors      |            29.802 |           30.994 |               32.902 |               38.862 |           689.030 |              6.794
 vector_concat | medium_vectors     |            36.001 |           37.909 |               45.061 |               49.829 |           238.180 |              3.398
 vector_concat | large_vectors      |           780.106 |          813.007 |              838.995 |              845.909 |          1189.947 |             14.815
 vector_concat | very_large_vectors |          2048.016 |         2074.003 |             2094.030 |             2108.812 |          3163.099 |             16.784
```