@binarycleric 2025-06-18
@jkatz Sure thing. Sorry about that.