@ankane 2025-06-18
This function isn't used for nearest neighbor search, so additional benchmarks shouldn't be needed. Will take a look at this after #860.