@binarycleric 2025-06-19
Sounds good. I'll update that in the morning.